chore: migrate helper functions to linux
