+++ src/otaclient_common/_io.py
@@ -25,7 +25,7 @@ from functools import partial
 from pathlib import Path
 
 from otaclient_common._typing import StrOrPath
-from otaclient_common.common import is_directory
+from otaclient_common.linux import is_directory
 
 logger = logging.getLogger(__name__)
