+++ src/otaclient_common/linux.py
@@ -17,6 +17,7 @@ from __future__ import annotations
 
 import os
 import shlex
+import stat
 import subprocess
 from pathlib import Path
 from subprocess import check_call
@@ -302,3 +303,34 @@ def fstrim_at_subprocess(
         stderr=subprocess.DEVNULL,
         shell=True,
     )
+
+
+def is_regular_file(path: Path) -> bool:
+    """Check if a file is regular."""
+    try:
+        path_stat = path.lstat()
+        return (
+            stat.S_ISREG(path_stat.st_mode)
+            and not stat.S_ISLNK(path_stat.st_mode)
+            and path_stat.st_size > 0
+        )
+    except OSError:
+        return False
+
+
+def is_file_or_symlink(path: Path) -> bool:
+    """Check if a file is regular or symlink."""
+    try:
+        path_stat = path.lstat()
+        return stat.S_ISREG(path_stat.st_mode) or stat.S_ISLNK(path_stat.st_mode)
+    except OSError:
+        return False
+
+
+def is_directory(path: Path) -> bool:
+    """Check if a directory is regular."""
+    try:
+        path_stat = path.lstat()
+        return stat.S_ISDIR(path_stat.st_mode) and not stat.S_ISLNK(path_stat.st_mode)
+    except OSError:
+        return False
